CultuCat_Backend/CultuCat_Django/requirements.txt
asgiref==3.8.1
CacheControl==0.12.6
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
cryptography==3.4.8
Django==3.2.20
djangorestframework==3.12.4
firebase-admin==5.3.0
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0.dev2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==2.10
msgpack==1.1.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==20.0.1
pyparsing==3.1.4
pytz==2025.2
requests==2.25.1
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.26.15